+++ TabletopTweaks/NewContent/Feats/ExtraReservoir.cs
@@ -1,13 +1,9 @@
 ﻿using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using TabletopTweaks.Config;
 using TabletopTweaks.Extensions;
 using TabletopTweaks.Utilities;
+using static Kingmaker.Blueprints.Classes.Prerequisites.Prerequisite;
 
 namespace TabletopTweaks.NewContent.Feats {
     static class ExtraReservoir {
@@ -20,7 +16,8 @@ namespace TabletopTweaks.NewContent.Feats {
                 bp.SetName("Extra Reservoir");
                 bp.SetDescription("You gain three more points in your arcane reservoir, and the maximum number of points in your arcane reservoir " +
                     "increases by that amount.\nYou can take this feat multiple times. Its effects stack.");
-                bp.AddPrerequisiteFeaturesFromList(1, ArcanistArcaneReservoirFeature, ArcaneEnforcerArcaneReservoirFeature);
+                bp.AddPrerequisiteFeature(ArcanistArcaneReservoirFeature, GroupType.Any);
+                bp.AddPrerequisiteFeature(ArcaneEnforcerArcaneReservoirFeature, GroupType.Any);
             });
             if (ModSettings.AddedContent.Feats.IsDisabled("ExtraReservoir")) { return; }
             FeatTools.AddAsFeat(ExtraReservoir);