+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Version 1.1.0
+* added extra reservoir
+
 ## Version 1.0.9
 * added focused weapon advanced weapon training
 * eldritch scion can now properly take a second bloodline